fix: change WMSLoader name to WMSProvider and fix identation
